% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stef_global_spatial_normaliser.R
\name{stef_global_spatial_normaliser}
\alias{stef_global_spatial_normaliser}
\title{Global spatial normalisation}
\usage{
stef_global_spatial_normaliser(inraster, isStack = T, xpercentile = 0.95,
  output_filename = NULL)
}
\arguments{
\item{inraster}{Input raster stack or single raster.}

\item{isStack}{Logical. Set to TRUE if the input raster is a raster stack.}

\item{xpercentile}{The upper  percentile to use for normalisation. Default is 0.95, which represent 95th percentile.}

\item{output_filename}{The name of the output normalised raster or raster stack. The name must contain the file format (e.g. raster_normalised.tif)}
}
\description{
Performs global spatial normalisation of satellite data to reduce seasonal variations and intersensor dfferences. For local spatial normalisation, see \code{\link[STEF]{stef_local_spatial_normaliser}}
}
\details{
To be completed.
}
\examples{
\dontrun{

#create a raster stack
ra <- raster(ncols=360, nrows=180)
ra[] <- rnorm(ncell(ra))
for (i in 1: 86){
ro <- raster(ncols=360, nrows=180)
ro[] <- rnorm(ncell(ro))
ra <- stack(ra, ro)
}

# example:
stef_global_spatial_normaliser(ra, isStack = T, xpercentile = 0.95,output_filename =""ra_global_normalised.tif)
}
}
\author{
Eliakim Hamunyela
}
\references{
1. Hamunyela, E., Verbesselt, J., Herold, M. (2016) Using spatial context to improve early detection of deforestationfrom Landsat time series. Remote Sensing of Environment,172, 126–138.
  \url{http://dx.doi.org/10.1016/j.rse.2015.11.006}

2. Hamunyela, E., Verbesselt, J.,de Bruin, S., Herold, M. (2016). Monitoring Deforestation at Sub-Annual Scales as Extreme Events in Landsat Data Cubes. Remote Sensing, 8(8), 651.
 \url{http://dx.doi.org/10.3390/rs8080651}

  3. Hamunyela, E., Reiche, J., Verbesselt, J., & Herold, M. (2017). Using space-time features to improve detection of forest disturbances from Landsat time series. Remote Sensing, 9(6), 515.
   \url{http://dx.doi:10.3390/rs9060515}
}

